compiling new backend. Removed Statistics and added to chart response
