+++ launcher-tool-go/pkg/launcher/config/im2k8s/config.yaml
@@ -66,6 +66,19 @@ services:
           unit: Cpu
           interval: Minutely
 
+      u2-standard:
+        nameSuffix: Cpu
+
+        cpu: [ 1, 2, 4 ]
+        memory: [ 1, 2, 4 ]
+
+        cpuModel: Model
+        memoryModel: Model
+
+        payment:
+          type: Resource
+          unit: Cpu
+          interval: Minutely
   fileSystem:
     name: storage
     mountPoint: /mnt/storage
